refactor(cli-runtime): split gateway client by responsibility
